requirements.txt
aiohttp==3.6.2
alembic==1.4.2
async-timeout==3.0.1
attrs==19.3.0
backcall==0.2.0
chardet==3.0.4
click==7.1.2
decorator==4.4.2
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
idna==2.9
idna-ssl==1.1.0
ipdb==0.13.3
ipython==7.16.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
mixer==6.1.3
multidict==4.7.6
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.6
ptyprocess==0.6.0
Pygments==2.6.1
pytest==6.0.1
pytest-clarity==0.3.0a0
pytest-flask==1.0.0
pytest-sugar==0.9.4
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
six==1.15.0
slackclient==2.8.0
slackeventsapi==2.2.1
SQLAlchemy==1.3.19
traitlets==4.3.3
typing-extensions==3.7.4.2
wcwidth==0.2.5
Werkzeug==1.0.1
yarl==1.4.2
